Fix time zone being sent as time and allow error transparency
A bad parameter passed to genResponsePayload was causing the time zone
name to be sent back in both response fields. Whoops!

Additionally, sending an error back short circuits any response you
might have set up and just gives the caller an "internal server error"
message. This makes sense, so the return call in badRequest was changed
to return nil as the error, to allow the custom response to be shown to
the caller.